Packaging only eggs or source
APY-142 : Rewrite packiging code

Added option for packaging only eggs or source distribution.

+++ distribution-manager/ally-distribution/ally/distribution/packaging/impl/processor/build.py
@@ -38,8 +38,8 @@ class Package(Context):
     '''
     # ---------------------------------------------------------------- Required
     path = requires(str)
-    pathSetupPy = requires(str)
     name = requires(str)
+    setup = requires(str)
 
 # --------------------------------------------------------------------
 
@@ -49,10 +49,13 @@ class BuildHandler(HandlerProcessor):
     Implementation for a processor that provides the package egg build.
     '''
     
+    cmds = list
+    # The setup util distribution commands.
     pathBuild = str
     # The location where the builds are placed. 
     
     def __init__(self):
+        assert isinstance(self.cmds, list), 'Invalid commands %s' % self.cmds
         assert isinstance(self.pathBuild, str), 'Invalid build path %s' % self.pathBuild
         super().__init__(Package=Package)
 
@@ -64,6 +67,7 @@ class BuildHandler(HandlerProcessor):
         '''
         assert isinstance(distribution, Distribution), 'Invalid distribution %s' % distribution
         if not distribution.packages: return
+        if not self.cmds: return
             
         savedCwd = os.getcwd()
              
@@ -72,13 +76,25 @@ class BuildHandler(HandlerProcessor):
 
             os.chdir(package.path)
             
+            with open('setup.py', 'w') as f: print(package.setup, file=f)
+            
             buildPath = os.path.join(savedCwd, self.pathBuild)
-            run_setup(package.pathSetupPy, ('sdist', '--dist-dir', buildPath, 'bdist_egg', '--dist-dir', buildPath))
-            # run_setup(package.pathSetupPy, ('--help', 'sdist'))
+            log.info('%s Building %s', '=' * 50, package.name)
+            
+            commands = []
+            for cmd in self.cmds:
+                commands.append(cmd)
+                commands.append('--dist-dir')
+                commands.append(buildPath)
+            
+            try: run_setup('setup.py', commands)
+            except: log.exception('Cannot publish \'%s\'', package.path)
             
             # Cleaning setup directories.
-            rmtree('build')
-            rmtree('%s.egg-info' % package.name.replace('-', '_'))
+            os.remove('setup.py')
+            if os.path.exists('build'): rmtree('build')
+            eggInfo = '%s.egg-info' % package.name.replace('-', '_')
+            if os.path.exists(eggInfo): rmtree(eggInfo)
             
         # Restoring environment. 
         os.chdir(savedCwd)
